Added show_tasks() function to shorten code

+++ kanban.py
@@ -8,6 +8,17 @@ def showmenu():
     print("\n")
     print(Fore.LIGHTMAGENTA_EX+"1.View Kanban\n2.Add task\n3.Move tasks\n4.Delete task\n5.Quit")
     
+def show_tasks(status):
+  command = f"select Taskname,Priority,Reportee,Assignee from Kanban_table where Status ='{status}' order by Priority"
+  cursor.execute(command)
+  data = cursor.fetchall()
+  if(data):
+      ptable = PrettyTable(["Taskname","Priority","Reportee","Assignee"])
+      for a,b,c,d in data:
+          ptable.add_row([a,b,c,d])
+      print(ptable)
+  else:
+      print("-----Empty-----")
 
 def isempty():
     cursor.execute("select * from Kanban_Table;")
@@ -33,29 +44,13 @@ def displaykanban():
             rows = int(input("How many tasks would you like to add?"))
             addrows(rows)
     else:
-        #Print entire Kanban table
+    
         print(Fore.RED+"________𝗧𝗢-𝗗𝗢________")
-        ptable = PrettyTable(["Taskname","Priority","Reportee","Assignee"])
-        cursor.execute("select Taskname,Priority,Reportee,Assignee from Kanban_table where Status ='To-do' order by Priority")
-        data = cursor.fetchall()
-        if(data):
-            for a,b,c,d in data:
-                ptable.add_row([a,b,c,d])
-            print(ptable)
-        else:
-            print("-----Empty-----")
+        show_tasks("To-Do")
         
 
         print(Fore.YELLOW+"________𝗜𝗡-𝗣𝗥𝗢𝗚𝗥𝗘𝗦𝗦________")
-        ptable = PrettyTable(["Taskname","Priority","Reportee","Assignee"])
-        cursor.execute("select Taskname,Priority,Reportee,Assignee from Kanban_table where Status ='In-progress' order by Priority")
-        data = cursor.fetchall()
-        if(data):
-            for a,b,c,d in data:
-                ptable.add_row([a,b,c,d])
-            print(ptable)
-        else:
-            print("-----Empty-----") 
+        show_tasks("In-Progress")
 
         print(Fore.GREEN+"________𝗗𝗢𝗡𝗘________")
         ptable = PrettyTable(["Taskname","Priority","Reportee","Assignee"])
@@ -69,18 +64,36 @@ def displaykanban():
             print("-----Empty-----")
 
 def movetask():
-    task = input("Enter task name: ")
-    movelocation = input("Enter new status of task (To-do,In-progress,Done): ")
-    cursor.execute("update kanban_table set status = '{}' where taskname = '{}'".format(movelocation,task))
-    mycon.commit()
+    if not isempty():
+        tasknames = PrettyTable(["Task names","Status"])
+        cursor.execute("select Taskname,Status from Kanban_table")
+        data = cursor.fetchall()
+        for a,b in data:
+            tasknames.add_row([a,b])
+        print(tasknames)
+        task = input("Enter task name: ")
+        movelocation = input("Enter new status of task (To-do,In-progress,Done): ")
+        cursor.execute("update kanban_table set status = '{}' where taskname = '{}'".format(movelocation,task))
+        mycon.commit()
+    else:
+        print("No tasks present")
 
 def deltask():
-    task = input("Enter task name: ")
-    confirm = input(Fore.RED+"Confirm deletion (y/n): ")
-    if (confirm=='y'):
-        cursor.execute("delete from kanban_table where taskname = '{}' ".format(task))
-        mycon.commit()
-        print(Fore.LIGHTYELLOW_EX)
+    if not isempty():
+        tasknames = PrettyTable(["Task names","Status"])
+        cursor.execute("select Taskname,Status from Kanban_table")
+        data = cursor.fetchall()
+        for a,b in data:
+            tasknames.add_row([a,b])
+        print(tasknames)
+        task = input("Enter task name: ")
+        confirm = input(Fore.RED+"Confirm deletion (y/n): ")
+        if (confirm=='y'):
+            cursor.execute("delete from kanban_table where taskname = '{}' ".format(task))
+            mycon.commit()
+            print(Fore.LIGHTYELLOW_EX)
+    else:
+        print("No tasks present")
 
 def reset():
     c = input(Fore.RED+"Confirm reset (y/n): ")
